shortened databasepreparation skip pdb_relaxed.
No relaxation needed at the step inbetween, so skipped this thing an modified dbgen.py, and preprocess.py. This is nessecary here becaus elfin-data will get this files when setting up from here!

ALSO this means if one wants to update it elfin-data is necessary to redo the setup because simple fetching won't solve the problem!